refactor(util/itinerary): Move helper function from state to itinerary util.
