+++ lib/util/state.js
@@ -5,15 +5,15 @@ import { FormattedList, FormattedMessage } from 'react-intl'
 import { isTransit } from '@opentripplanner/core-utils/lib/itinerary'
 import { itineraryToTransitive } from '@opentripplanner/transitive-overlay'
 import coreUtils from '@opentripplanner/core-utils'
-import hash from 'object-hash'
 import isEqual from 'lodash.isequal'
-import memoize from 'lodash.memoize'
 import qs from 'qs'
 import React from 'react'
 import styled from 'styled-components'
 
 import FormattedMode from '../components/util/formatted-mode'
 import Strong from '../components/util/strong-text'
+
+import { collectItinerariesWithoutDuplicates } from './itinerary'
 
 // For lowercase context
 const LowerCase = styled.span`
@@ -253,33 +253,6 @@ export const getActiveFieldTripRequest = createSelector(
   }
 )
 
-// Ignore certain keys that could add significant calculation time to hashing.
-// The alerts are irrelevant, but the intermediateStops, interStopGeometry and
-// steps could have the legGeometry substitute as an equivalent hash value
-const blackListedKeys = [
-  'alerts',
-  'intermediateStops',
-  'interStopGeometry',
-  'steps'
-]
-// make blackListedKeys into an object due to superior lookup performance
-const blackListedKeyLookup = {}
-blackListedKeys.forEach((key) => {
-  blackListedKeyLookup[key] = true
-})
-
-/**
- * A memoized function to hash the itinerary.
- * NOTE: It can take a while (>30ms) for the object-hash library to calculate
- * an itinerary's hash for some lengthy itineraries. If better performance is
- * desired, additional values to blackListedKeys should be added to avoid
- * spending extra time hashing values that wouldn't result in different
- * itineraries.
- */
-const hashItinerary = memoize((itinerary) =>
-  hash(itinerary, { excludeKeys: (key) => blackListedKeyLookup[key] })
-)
-
 /**
  * Get the active itineraries for the active search, which is dependent on
  * whether realtime or non-realtime results should be displayed
@@ -301,24 +274,8 @@ export const getActiveItineraries = createSelector(
     activeFieldTripRequest
   ) => {
     const response = realtimeResponse || nonRealtimeResponse
-    const itineraries = []
-    // keep track of itinerary hashes in order to not include duplicate
-    // itineraries. Duplicate itineraries can occur in batch routing where a walk
-    // to transit trip can sometimes still be the most optimal trip even when
-    // additional modes such as bike rental were also requested
-    const seenItineraryHashes = {}
-    if (response) {
-      response.forEach((res) => {
-        res?.plan?.itineraries?.forEach((itinerary) => {
-          // hashing takes a while on itineraries
-          const itineraryHash = hashItinerary(itinerary)
-          if (!seenItineraryHashes[itineraryHash]) {
-            itineraries.push(itinerary)
-            seenItineraryHashes[itineraryHash] = true
-          }
-        })
-      })
-    }
+    const itineraries = collectItinerariesWithoutDuplicates(response)
+
     const { sort } = itinerarySortSettings
     const { direction, type } = sort
     // If no sort type is provided (e.g., because batch routing is not enabled),
