+++ package.json
@@ -138,7 +138,9 @@
     "@percy/puppeteer": "^2.0.2",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
     "@types/clone": "^2.1.1",
+    "@types/lodash.memoize": "^4.1.8",
     "@types/mapbox__polyline": "^1.0.2",
+    "@types/object-hash": "^3.0.5",
     "@types/qs": "^6.9.7",
     "@types/react-bootstrap": "^0.32.26",
     "@types/react-router": "^5.1.17",
